+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <cpp_log/cpp_log_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  CppLogPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("CppLogPluginCApi"));
 }